remove unused highlightTermsInTextTag, extract buildResultFromMatches into its own file
