+++ packages/react-tei/src/termEnrichment/types.ts
@@ -1,3 +1,5 @@
+import type { DocumentJson } from "../parser/document";
+
 export type NormalizedTerm = {
 	term: string;
 	group: string;
@@ -29,4 +31,33 @@ export type SubTermAtPosition = {
 	start: number;
 	end: number;
 	fromTermGroups: string[];
+};
+
+export type TextTag = {
+	tag: "#text";
+	attributes?: Record<string, string>;
+	value: string;
+} & DocumentJson;
+
+export type HighlightTag = {
+	tag: "highlight";
+	attributes: {
+		groups: string[];
+		term: string;
+		noAnchor?: boolean;
+	};
+	value: (TextTag | HighlightTag)[];
+} & DocumentJson;
+
+export type HighlightedTextTag = {
+	tag: "highlightedText";
+	attributes?: Record<string, string>;
+	value: (HighlightTag | TextTag)[];
+} & DocumentJson;
+
+export type TermData = {
+	termRegex: RegExp;
+	term: string;
+	groups: string[];
+	value?: (TextTag | HighlightTag)[] | string;
 };
